moving attachments to contacts and adding better logging support
